backend/pyproject.toml
[project]
name = "kis-backend"
version = "0.1.0"
description = "KIS Django REST API Backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "djangorestframework-simplejwt>=5.3",
    "psycopg[binary]>=3.1",
    "python-dotenv>=1.0",
    "gunicorn>=21.0",
    "django-cors-headers>=4.3",
    "dj-database-url>=2.1",
]

[dependency-groups]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.5",
]
